日程管理/mode；modeExchangeType

+++ BabySharing/Controller/Home/FilterController/SearchFilterDateController/CalendarView/AYDayCollectionCellView/AYDayCollectionCellView.h
@@ -14,7 +14,7 @@
 @property (nonatomic, copy) NSString *gregoiainDay;
 @property (nonatomic, copy) NSString *dayDay;
 @property (nonatomic, assign) BOOL isGone;
-@property (nonatomic, assign) long timeSpan;
+@property (nonatomic, assign) double timeSpan;
 /** 日 */
 @property (nonatomic, copy) NSString *lunarDay;
